billing_service/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-stripe==2.6.0
Django==3.2.16
idna==3.4
jsonfield==3.1.0
psycopg2-binary==2.9.5
requests==2.28.2
sqlparse==0.4.3
stripe==4.2.0
urllib3==1.26.15
confluent-kafka==1.9.0
python-dotenv==1.0.0
pyjwt==2.6.0
uwsgi==2.0.20